Fix circular deps by stripping type and dict
