+++ components/blog/article-parts/HighlightedBoxes.tsx
@@ -2,8 +2,8 @@ import styles from "../../../styles/blog/blog.module.scss";
 import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
 import {faFrog, faTriangleExclamation} from "@fortawesome/free-solid-svg-icons";
 import React from "react";
-import {ArticleParts} from "../../../lib/blog/articleParts";
 import ArticleRendererFromContext from "../../../app/blog/renderer/ArticleRenderer";
+import {ArticleParts} from "../ArticleParts";
 
 export const Caution: ArticleParts = content => (
   <div className={styles.caution}>
