+++ components/blog/article-parts/TwitterArchive.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
 import {faDove} from "@fortawesome/free-solid-svg-icons";
 import getOtakuColor from "../../../lib/blog/otakuColors";
-import {ArticleParts} from "../../../lib/blog/articleParts";
+import {ArticleParts} from "../ArticleParts";
 
 const TwitterArchive: ArticleParts = content => {
   const tweetData: { [key: string]: string } = {}
